rename files and small refactor
