+++ app/screens/post_options/components/quick_reactions/quick_reaction.tsx
@@ -1,43 +1,20 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {withDatabase} from '@nozbe/watermelondb/DatabaseProvider';
-import withObservables from '@nozbe/with-observables';
 import React, {useCallback} from 'react';
 import {useIntl} from 'react-intl';
 import {View} from 'react-native';
-import {of as of$} from 'rxjs';
-import {catchError, switchMap} from 'rxjs/operators';
 
 import CompassIcon from '@components/compass_icon';
-import {MM_TABLES, SYSTEM_IDENTIFIERS} from '@constants/database';
 import {EMOJI_PICKER} from '@constants/screens';
 import {dismissModal, showModal} from '@screens/navigation';
-import {safeParseJSON} from '@utils/helpers';
 
 import PickReaction from './pick_reaction';
 import Reaction from './reaction';
 
-import type {WithDatabaseArgs} from '@typings/database/database';
-import type SystemModel from '@typings/database/models/servers/system';
-
 type QuickReactionProps = {
     theme: Theme;
     recentEmojis: string[];
-};
-
-const DEFAULT_EMOJIS = [
-    'thumbsup',
-    'smiley',
-    'white_check_mark',
-    'heart',
-    'eyes',
-    'raised_hands',
-];
-
-const mergeRecentWithDefault = (recentEmojis: string[]) => {
-    const filterUsed = DEFAULT_EMOJIS.filter((e) => !recentEmojis.includes(e));
-    return recentEmojis.concat(filterUsed).splice(0, 6);
 };
 
 const QuickReaction = ({recentEmojis = [], theme}: QuickReactionProps) => {
@@ -103,14 +80,4 @@ const QuickReaction = ({recentEmojis = [], theme}: QuickReactionProps) => {
     );
 };
 
-const enhanced = withObservables([], ({database}: WithDatabaseArgs) => ({
-    recentEmojis: database.
-        get<SystemModel>(MM_TABLES.SERVER.SYSTEM).
-        findAndObserve(SYSTEM_IDENTIFIERS.RECENT_REACTIONS).
-        pipe(
-            switchMap((recent) => of$(mergeRecentWithDefault(safeParseJSON(recent.value) as string[]))),
-            catchError(() => of$(mergeRecentWithDefault([]))),
-        ),
-}));
-
-export default withDatabase(enhanced(QuickReaction));
+export default QuickReaction;
